requirements.txt
streamlit
plotly-express
selenium
tabula-py